fixed block unnest on backspace click
